std::date: consolidate the unit tests into the canonical __tests__/date.test.ts
There was already a date test suite at lib/stdlib/__tests__/date.test.ts
covering the old string API; the migration broke its 26 assertions (removed
add* helpers, now-as-string, atTime/startOf* as strings). Rewrote it for the
numbers API and folded in the new coverage (ms round-trip, DST spring/fall,
the endOf* ordering invariant, cross-month weeks, leap February, plus an
atTime case for the midnight-crossing offset the old code masked). Removed the
duplicate lib/stdlib/date.test.ts I had added. lib/stdlib now: 1224 pass, 0 fail.
